feat: load storage providers at strapi 'register' step
