+++ packages/core/upload/server/bootstrap.js
@@ -1,12 +1,8 @@
 'use strict';
 
-const _ = require('lodash');
-
 module.exports = async ({ strapi }) => {
   // set plugin store
   const configurator = strapi.store({ type: 'plugin', name: 'upload', key: 'settings' });
-
-  strapi.plugin('upload').provider = createProvider(strapi.config.get('plugin.upload', {}));
 
   // if provider config does not exist set one by default
   const config = await configurator.get();
@@ -22,62 +18,6 @@ module.exports = async ({ strapi }) => {
   }
 
   await registerPermissionActions();
-};
-
-const createProvider = config => {
-  const { providerOptions, actionOptions = {} } = config;
-
-  const providerName = _.toLower(config.provider);
-  let provider;
-
-  let modulePath;
-  try {
-    modulePath = require.resolve(`@strapi/provider-upload-${providerName}`);
-  } catch (error) {
-    if (error.code === 'MODULE_NOT_FOUND') {
-      modulePath = providerName;
-    } else {
-      throw error;
-    }
-  }
-
-  try {
-    provider = require(modulePath);
-  } catch (err) {
-    throw new Error(`Could not load upload provider "${providerName}".`);
-  }
-
-  const providerInstance = provider.init(providerOptions);
-
-  if (!providerInstance.delete) {
-    throw new Error(`The upload provider "${providerName}" doesn't implement the delete method.`);
-  }
-
-  if (!providerInstance.upload && !providerInstance.uploadStream) {
-    throw new Error(
-      `The upload provider "${providerName}" doesn't implement the uploadStream nor the upload method.`
-    );
-  }
-
-  if (!providerInstance.uploadStream) {
-    process.emitWarning(
-      `The upload provider "${providerName}" doesn't implement the uploadStream function. Strapi will fallback on the upload method. Some performance issues may occur.`
-    );
-  }
-
-  const wrappedProvider = _.mapValues(providerInstance, (method, methodName) => {
-    return async function(file, options = actionOptions[methodName]) {
-      return providerInstance[methodName](file, options);
-    };
-  });
-
-  return Object.assign(Object.create(baseProvider), wrappedProvider);
-};
-
-const baseProvider = {
-  extend(obj) {
-    Object.assign(this, obj);
-  },
 };
 
 const registerPermissionActions = async () => {
