+++ packages/core/upload/server/register.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const _ = require('lodash');
 const registerUploadMiddleware = require('./middlewares/upload');
 
 /**
@@ -7,9 +8,67 @@ const registerUploadMiddleware = require('./middlewares/upload');
  * @param {{ strapi: import('@strapi/strapi').Strapi }}
  */
 module.exports = async ({ strapi }) => {
+  strapi.plugin('upload').provider = createProvider(strapi.config.get('plugin.upload', {}));
+
   await registerUploadMiddleware({ strapi });
 
   if (strapi.plugin('graphql')) {
     require('./graphql')({ strapi });
   }
+};
+
+const createProvider = (config) => {
+  const { providerOptions, actionOptions = {} } = config;
+
+  const providerName = _.toLower(config.provider);
+  let provider;
+
+  let modulePath;
+  try {
+    modulePath = require.resolve(`@strapi/provider-upload-${providerName}`);
+  } catch (error) {
+    if (error.code === 'MODULE_NOT_FOUND') {
+      modulePath = providerName;
+    } else {
+      throw error;
+    }
+  }
+
+  try {
+    provider = require(modulePath);
+  } catch (err) {
+    throw new Error(`Could not load upload provider "${providerName}".`);
+  }
+
+  const providerInstance = provider.init(providerOptions);
+
+  if (!providerInstance.delete) {
+    throw new Error(`The upload provider "${providerName}" doesn't implement the delete method.`);
+  }
+
+  if (!providerInstance.upload && !providerInstance.uploadStream) {
+    throw new Error(
+      `The upload provider "${providerName}" doesn't implement the uploadStream nor the upload method.`
+    );
+  }
+
+  if (!providerInstance.uploadStream) {
+    process.emitWarning(
+      `The upload provider "${providerName}" doesn't implement the uploadStream function. Strapi will fallback on the upload method. Some performance issues may occur.`
+    );
+  }
+
+  const wrappedProvider = _.mapValues(providerInstance, (method, methodName) => {
+    return async function (file, options = actionOptions[methodName]) {
+      return providerInstance[methodName](file, options);
+    };
+  });
+
+  return Object.assign(Object.create(baseProvider), wrappedProvider);
+};
+
+const baseProvider = {
+  extend(obj) {
+    Object.assign(this, obj);
+  },
 };
